Move rocket launch systems/components to module SOL-12 #done
